# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Webinar platform.",
        "title": "Webinar Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/invites": {
            "post": {
                "security": [
                    {
                        "KratosSession": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "Create an invite",
                "parameters": [
                    {
                        "description": "Invite details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organization.CreateInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/organization.Invite"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invites/pending": {
            "get": {
                "security": [
                    {
                        "KratosSession": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "List pending invites",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/organization.Invite"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "organization.CreateInviteRequest": {
            "type": "object",
            "required": [
                "email",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "organization.Invite": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invited_by": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/organization.Role"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "organization.Role": {
            "type": "string",
            "enum": [
                "admin",
                "manager",
                "human-resource",
                "developer"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleManager",
                "RoleHumanResource",
                "RoleDeveloper"
            ]
        },
...
    },
    "securityDefinitions": {
        "KratosSession": {
            "description": "Ory Kratos session cookie obtained via the self-service login flow.",
            "type": "apiKey",
            "name": "ory_kratos_session",
            "in": "cookie"
        }
    }
}