# sqlc/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Invite struct {
	ID             uuid.UUID
	OrganizationID uuid.UUID
	Email          string
	Role           string
	InvitedBy      uuid.UUID
	Accepted       bool
	ExpiresAt      pgtype.Timestamptz
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
}

type Organization struct {
	ID        uuid.UUID
	Name      string
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type User struct {
	ID             uuid.UUID
	IdentityID     uuid.UUID
	OrganizationID pgtype.UUID
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
}